+++ src/features/comments/lib/queryKey.ts
@@ -1,6 +1,6 @@
 import { createQueryKeys } from "@lukemorales/query-key-factory";
 
-import { Comment } from "@/entities/comments/model";
+import { Comment } from "@/entities/comments";
 
 export const commentsKeys = createQueryKeys("comments", {
   getComments: (postId: Comment["postId"]) => ["all", postId],
