+++ src/features/users/lib/queryKey.ts
@@ -1,6 +1,6 @@
 import { createQueryKeys } from "@lukemorales/query-key-factory";
 
-import { User } from "@/entities/users/model";
+import { User } from "@/entities/users";
 
 export const usersKeys = createQueryKeys("users", {
   getUsers: () => ["all"],
